use wrf in run_iRF
