+++ irf/ensemble/__init__.py
@@ -1,3 +1,3 @@
 from .wrf import (RandomForestClassifierWithWeights, 
                   RandomForestRegressorWithWeights)
-from .wrf import wrf
+from .wrf import (wrf, wrf_reg)
