+++ irf/ensemble/wrf.py
@@ -4,6 +4,7 @@ from sklearn.base import clone
 from abc import ABCMeta, abstractmethod
 from ..tree.tree import (WeightedDecisionTreeClassifier, 
                          WeightedDecisionTreeRegressor)
+from ..utils import get_rf_tree_data
 
 class RandomForestClassifierWithWeights(RandomForestClassifier):
     def fit(self, X, y, sample_weight=None, feature_weight=None):
@@ -22,9 +23,13 @@ class RandomForestRegressorWithWeights(RandomForestRegressor):
         return super(RandomForestRegressorWithWeights, self).fit(X, y, sample_weight)
         
 class wrf(RandomForestClassifierWithWeights):
-    def fit(self, X, y, sample_weight=None, feature_weight=None, K = 5, keep_rf = True):
+    def fit(self, X, y, sample_weight=None, feature_weight=None, K = 5, 
+            keep_record = True, X_test = None, y_test = None):
         self.all_rf_weights = dict()
-        self.all_rfs = dict()
+        if keep_record:
+            self.all_K_iter_rf_data = dict()
+            assert X_test is not None, 'X_test should not be None when keep_record'
+            assert y_test is not None, 'y_test should not be None when keep_record'
         for k in range(K):
             if k == 0:
                 # Initially feature weights are None
@@ -42,17 +47,24 @@ class wrf(RandomForestClassifierWithWeights):
             # new feature importance score
             feature_importances = self.feature_importances_
             self.all_rf_weights["rf_weight{}".format(k + 1)] = feature_importances
-            if keep_rf:
-                self.all_rfs['rf_{}'.format(k+1)] = clone(self)
-                self.all_rfs['rf_{}'.format(k+1)].estimators_ = clone(self.estimators_)
+            if keep_record:
+                self.all_K_iter_rf_data["rf_iter{}".format(k+1)] = get_rf_tree_data(
+                        rf=self,
+                        X_train=X,
+                        X_test=X_test,
+                        y_test=y_test)
         return self
 
 #Eric: Doesn't lfook like much is changed here, as it looks like no significant differences between regressor and classifier
 #       come back to later to make sure there really is no difference                
 class wrf_reg(RandomForestRegressorWithWeights): # Hue: change the name so that it does not clash with the first one.
-    def fit(self, X, y, sample_weight=None, feature_weight=None, K = 5, keep_rf = True):
+    def fit(self, X, y, sample_weight=None, feature_weight=None, K = 5, 
+            keep_record = True, X_test = None, y_test = None):
         self.all_rf_weights = dict()
-        self.all_rfs = dict()
+        if keep_record:
+            self.all_K_iter_rf_data = dict()
+            assert X_test is not None, 'X_test should not be None when keep_record'
+            assert y_test is not None, 'y_test should not be None when keep_record'
         for k in range(K):
             if k == 0:
                 # Initially feature weights are None
@@ -70,7 +82,10 @@ class wrf_reg(RandomForestRegressorWithWeights): # Hue: change the name so that 
             # new feature importance score
             feature_importances = self.feature_importances_
             self.all_rf_weights["rf_weight{}".format(k + 1)] = feature_importances
-            if keep_rf:
-                self.all_rfs['rf_{}'.format(k+1)] = clone(self)
-                self.all_rfs['rf_{}'.format(k+1)].estimators_ = clone(self.estimators_)
+            if keep_record:
+                self.all_K_iter_rf_data["rf_iter{}".format(k+1)] = get_rf_tree_data(
+                        rf=self,
+                        X_train=X_train,
+                        X_test=X_test,
+                        y_test=y_test)
         return self
