+++ irf/irf_utils.py
@@ -8,456 +8,19 @@ from functools import partial
 from functools import reduce
 from scipy import stats
 import matplotlib.pyplot as plt
+from sklearn.base import (clone, 
+                          ClassifierMixin,
+                          RegressorMixin)
+from .utils import get_rf_tree_data
 
 # Needed for the scikit-learn wrapper function
 from sklearn.utils import resample
-from .ensemble import RandomForestClassifierWithWeights
+from .ensemble import (RandomForestClassifierWithWeights,
+                       RandomForestRegressorWithWeights,
+                       wrf, wrf_reg)
+
 from math import ceil
 
-
-def all_tree_paths(dtree, root_node_id=0):
-    """
-    Get all the individual tree paths from root node to the leaves
-    for a decision tree classifier object [1]_.
-
-    Parameters
-    ----------
-    dtree : DecisionTreeClassifier object
-        An individual decision tree classifier object generated from a
-        fitted RandomForestClassifierWithWeights object in scikit learn.
-
-    root_node_id : int, optional (default=0)
-        The index of the root node of the tree. Should be set as default to
-        0 and not changed by the user
-
-    Returns
-    -------
-    paths : list
-        Return a list containing 1d numpy arrays of the node paths
-        taken from the root node to the leaf in the decsion tree
-        classifier. There is an individual array for each
-        leaf node in the decision tree.
-
-    Notes
-    -----
-        To obtain a deterministic behaviour during fitting,
-        ``random_state`` has to be fixed.
-
-    References
-    ----------
-        .. [1] https://en.wikipedia.org/wiki/Decision_tree_learning
-
-    Examples
-    --------
-    >>> from sklearn.datasets import load_breast_cancer
-    >>> from sklearn.model_selection import train_test_split
-    >>> from irf.ensemble import RandomForestClassifierWithWeights
-    >>> raw_data = load_breast_cancer()
-    >>> X_train, X_test, y_train, y_test = train_test_split(
-        raw_data.data, raw_data.target, train_size=0.9,
-        random_state=2017)
-    >>> rf = RandomForestClassifierWithWeights(
-        n_estimators=3, random_state=random_state_classifier)
-    >>> rf.fit(X=X_train, y=y_train)
-    >>> estimator0 = rf.estimators_[0]
-    >>> tree_dat0 = getTreeData(X_train = X_train,
-                                dtree = estimator0,
-                                root_node_id = 0)
-    >>> tree_dat0['all_leaf_node_classes']
-    ...                             # doctest: +SKIP
-    ...
-    [1, 0, 1, 0, 1, 0, 1, 1, 0, 1, 1, 0, 0, 1, 0, 1, 1, 1, 0, 1, 0, 0]
-    """
-
-    # Use these lists to parse the tree structure
-    children_left = dtree.tree_.children_left
-    children_right = dtree.tree_.children_right
-
-    if root_node_id is None:
-        paths = []
-
-    if root_node_id == _tree.TREE_LEAF:
-        raise ValueError("Invalid node_id %s" % _tree.TREE_LEAF)
-
-    # if left/right is None we'll get empty list anyway
-    if children_left[root_node_id] != _tree.TREE_LEAF:
-        paths = [np.append(root_node_id, l)
-                 for l in all_tree_paths(dtree, children_left[root_node_id]) +
-                 all_tree_paths(dtree, children_right[root_node_id])]
-
-    else:
-        paths = [root_node_id]
-    return paths
-
-
-def get_validation_metrics(inp_class_reg_obj, y_true, X_test):
-    """
-    Get the various Random Forest/ Decision Tree metrics
-    This is currently setup only for classification forests and trees
-        TODO/ CHECK: We need to update this for regression purposes later
-        TODO/ CHECK: For classification we need to validate that
-               the maximum number of
-               labels is 2 for the training/ testing data
-
-    Get all the individual tree paths from root node to the leaves
-    for a decision tree classifier object [1]_.
-
-    Parameters
-    ----------
-    inp_class_reg_obj : DecisionTreeClassifier or RandomForestClassifierWithWeights
-        object [1]_
-        An individual decision tree or random forest classifier
-        object generated from a fitted Classifier object in scikit learn.
-
-    y_true : 1d array-like, or label indicator array / sparse matrix
-        Ground truth (correct) target values.
-
-    y_pred : 1d array-like, or label indicator array / sparse matrix
-        Estimated targets as returned by a classifier.
-
-    Returns
-    -------
-    classification_metrics : dict
-        Return a dictionary containing various validation metrics on
-        the input fitted Classifier object
-
-    References
-    ----------
-        .. [1] https://en.wikipedia.org/wiki/Decision_tree_learning
-
-    Examples
-    --------
-    >>> from sklearn.datasets import load_breast_cancer
-    >>> from sklearn.model_selection import train_test_split
-    >>> from irf.ensemble import RandomForestClassifierWithWeights
-    >>> raw_data = load_breast_cancer()
-    >>> X_train, X_test, y_train, y_test = train_test_split(
-        raw_data.data, raw_data.target, train_size=0.9,
-        random_state=2017)
-    >>> rf = RandomForestClassifierWithWeights(
-        n_estimators=3, random_state=random_state_classifier)
-    >>> rf.fit(X=X_train, y=y_train)
-    >>> rf_metrics = get_validation_metrics(inp_class_reg_obj = rf,
-                                          y_true = y_test,
-                                          X_test = X_test)
-    >>> rf_metrics['confusion_matrix']
-
-    ...                             # doctest: +SKIP
-    ...
-    array([[12,  2],
-          [ 1, 42]])
-    """
-
-    # If the object is not a scikit learn classifier then let user know
-    if type(inp_class_reg_obj).__name__ not in \
-       ["DecisionTreeClassifier", "WeightedDecisionTreeClassifier", "RandomForestClassifierWithWeights"]:
-        raise TypeError("input needs to be a DecisionTreeClassifier object, \
-        you have input a {} object".format(type(inp_class_reg_obj)))
-
-    # if the number of classes is not binary let the user know accordingly
-    if inp_class_reg_obj.n_classes_ != 2:
-        raise ValueError("The number of classes for classification must \
-        be binary, you currently have fit to {} \
-        classes".format(inp_class_reg_obj.n_classes_))
-
-    # Get the predicted values on the validation data
-    y_pred = inp_class_reg_obj.predict(X=X_test)
-
-    # CLASSIFICATION metrics calculations
-
-    # Cohen's kappa: a statistic that measures inter-annotator agreement.
-    # cohen_kappa_score = metrics.cohen_kappa_score(y1, y2[, labels, ...])
-
-    # Compute Area Under the Curve (AUC) using the trapezoidal rule
-    # fpr, tpr, thresholds = metrics.roc_curve(y_true = y_true,
-    #                                          y_pred = y_pred)
-    # auc = metrics.auc(fpr, tpr)
-
-    # Compute average precision (AP) from prediction scores
-    # average_precision_score = metrics.average_precision_score(y_true =
-    # y_true, y_score)
-
-    # Compute the Brier score.
-    # metrics.brier_score_loss(y_true = y_true, y_prob[, ...])
-
-    # Compute the F-beta score
-    # metrics.fbeta_score(y_true = y_true, y_pred = y_pred, beta[, ...])
-
-    # Average hinge loss (non-regularized)
-    # metrics.hinge_loss(y_true = y_true, pred_decision[, ...])
-
-    # Compute the Matthews correlation coefficient (MCC) for binary classes
-    # metrics.matthews_corrcoef(y_true = y_true, y_pred[, ...])
-
-    # Compute precision-recall pairs for different probability thresholds
-    # metrics.precision_recall_curve(y_true = y_true, ...)
-
-    # Compute precision, recall, F-measure and support for each class
-    # metrics.precision_recall_fscore_support(...)
-
-    # Compute Area Under the Curve (AUC) from prediction scores
-    # metrics.roc_auc_score(y_true = y_true, y_score[, ...])
-
-    # Compute Receiver operating characteristic (ROC)
-    # metrics.roc_curve(y_true = y_true, y_score[, ...])
-
-    # Jaccard similarity coefficient score
-    # jaccard_similarity_score =
-    # metrics.jaccard_similarity_score(y_true = y_true, y_pred = y_pred)
-
-    # Compute the F1 score, also known as balanced F-score or F-measure
-    f1_score = metrics.f1_score(y_true=y_true, y_pred=y_pred)
-
-    # Compute the average Hamming loss.
-    hamming_loss = metrics.hamming_loss(y_true=y_true, y_pred=y_pred)
-
-    # Log loss, aka logistic loss or cross-entropy loss.
-    log_loss = metrics.log_loss(y_true=y_true, y_pred=y_pred)
-
-    # Compute the precision
-    precision_score = metrics.precision_score(y_true=y_true, y_pred=y_pred)
-
-    # Compute the recall
-    recall_score = metrics.recall_score(y_true=y_true, y_pred=y_pred)
-
-    # Accuracy classification score
-    accuracy_score = metrics.accuracy_score(y_true=y_true, y_pred=y_pred)
-
-    # Build a text report showing the main classification metrics
-    # classification_report = metrics.classification_report(
-    # y_true=y_true, y_pred=y_pred)
-
-    # Compute confusion matrix to evaluate the accuracy of a classification
-    confusion_matrix = metrics.confusion_matrix(y_true=y_true, y_pred=y_pred)
-
-    # Zero-one classification loss.
-    zero_one_loss = metrics.zero_one_loss(y_true=y_true, y_pred=y_pred)
-
-    # Load all metrics into a single dictionary
-    classification_metrics = {"hamming_loss": hamming_loss,
-                              "log_loss": log_loss,
-                              "recall_score": recall_score,
-                              "precision_score": precision_score,
-                              "accuracy_score": accuracy_score,
-                              "f1_score": f1_score,
-                              # "classification_report": classification_report,
-                              "confusion_matrix": confusion_matrix,
-                              "zero_one_loss": zero_one_loss}
-
-    return classification_metrics
-
-
-def get_tree_data(X_train, X_test, y_test, dtree, root_node_id=0):
-    """
-    This returns all of the required summary results from an
-    individual decision tree
-
-    Parameters
-    ----------
-    dtree : DecisionTreeClassifier object
-        An individual decision tree classifier object generated from a
-        fitted RandomForestClassifierWithWeights object in scikit learn.
-
-    X_train : array-like or sparse matrix, shape = [n_samples, n_features]
-        Training vector, where n_samples in the number of samples and
-        n_features is the number of features.
-
-    X_test : array-like or sparse matrix, shape = [n_samples, n_features]
-        Test vector, where n_samples in the number of samples and
-        n_features is the number of features.
-
-    y_test : 1d array-like, or label indicator array / sparse matrix
-        Ground truth (correct) target values.
-
-    root_node_id : int, optional (default=0)
-        The index of the root node of the tree. Should be set as default to
-        0 and not changed by the user
-
-    Returns
-    -------
-    tree_data : dict
-        Return a dictionary containing various tree metrics
-    from the input fitted Classifier object
-
-    Examples
-    --------
-    >>> from sklearn.datasets import load_breast_cancer
-    >>> from sklearn.model_selection import train_test_split
-    >>> from sklearn.ensemble import RandomForestClassifierWithWeights
-    >>> raw_data = load_breast_cancer()
-    >>> X_train, X_test, y_train, y_test = train_test_split(
-        raw_data.data, raw_data.target, train_size=0.9,
-        random_state=2017)
-    >>> rf = RandomForestClassifierWithWeights(
-        n_estimators=3, random_state=2018)
-    >>> rf.fit(X=X_train, y=y_train)
-    >>> estimator0 = rf.estimators_[0]
-    >>> estimator0_out = get_tree_data(X_train=X_train,
-                                     dtree=estimator0,
-                                     root_node_id=0)
-    >>> print(estimator0_out['all_leaf_nodes'])
-    ...                             # doctest: +SKIP
-    ...
-    [6, 8, 9, 10, 12, 14, 15, 19, 22, 23, 24, \
-     25, 26, 29, 30, 32, 34, 36, 37, 40, 41, 42]
-    """
-
-    max_node_depth = dtree.tree_.max_depth
-    n_nodes = dtree.tree_.node_count
-    value = dtree.tree_.value
-    n_node_samples = dtree.tree_.n_node_samples
-    root_n_node_samples = float(dtree.tree_.n_node_samples[0])
-    X_train_n_samples = X_train.shape[0]
-
-    # Get the total number of features in the training data
-    tot_num_features = X_train.shape[1]
-
-    # Get indices for all the features used - 0 indexed and ranging
-    # to the total number of possible features in the training data
-    all_features_idx = np.array(range(tot_num_features), dtype='int64')
-
-    # Get the raw node feature indices from the decision tree classifier
-    # attribute positive and negative - we want only non-negative indices
-    # It is hard to tell which features this came from i.e. indices
-    # are zero, corresponding feature columns for consistency
-    # in reference
-    node_features_raw_idx = dtree.tree_.feature
-
-    # Get the refined non-negative feature indices for each node
-    # Start with a range over the total number of features and
-    # subset the relevant indices from the raw indices array
-    # print(np.array(node_features_raw_idx))
-    node_features_idx = all_features_idx[np.array(node_features_raw_idx)]
-
-    # Count the unique number of features used
-    num_features_used = (np.unique(node_features_idx)).shape[0]
-
-    # Get all of the paths used in the tree
-    all_leaf_node_paths = all_tree_paths(dtree=dtree,
-                                         root_node_id=root_node_id)
-
-    # Get list of leaf nodes
-    # In all paths it is the final node value
-    all_leaf_nodes = [path[-1] for path in all_leaf_node_paths]
-
-    # Get the total number of training samples used in each leaf node
-    all_leaf_node_samples = [n_node_samples[node_id].astype(int)
-                             for node_id in all_leaf_nodes]
-
-    # Get proportion of training samples used in each leaf node
-    # compared to the training samples used in the root node
-    all_leaf_node_samples_percent = [
-        100. * n_leaf_node_samples / root_n_node_samples
-        for n_leaf_node_samples in all_leaf_node_samples]
-
-    # Final predicted values in each class at each leaf node
-    all_leaf_node_values = [value[node_id].astype(
-        int) for node_id in all_leaf_nodes]
-
-    # Scaled values of the leaf nodes in each of the binary classes
-    all_scaled_leaf_node_values = [value / X_train_n_samples
-                                   for value in all_leaf_node_values]
-
-    # Total number of training samples used in the prediction of
-    # each class at each leaf node
-    tot_leaf_node_values = [np.sum(leaf_node_values)
-                            for leaf_node_values in all_leaf_node_values]
-
-    # All leaf node depths
-    # The depth is 0 indexed i.e. root node has depth 0
-    leaf_nodes_depths = [np.size(path) - 1 for path in all_leaf_node_paths]
-
-    # Predicted Classes
-    # Check that we correctly account for ties in determining the class here
-    all_leaf_node_classes = [all_features_idx[np.argmax(
-        value)] for value in all_leaf_node_values]
-
-    # Get all of the features used along the leaf node paths i.e.
-    # features used to split a node
-    # CHECK: Why does the leaf node have a feature associated with it?
-    # Investigate further
-    # Removed the final leaf node value so that this feature does not get
-    # included currently
-    all_leaf_paths_features = [node_features_idx[path[:-1]]
-                               for path in all_leaf_node_paths]
-
-    # Get the unique list of features along a path
-    # NOTE: This removes the original ordering of the features along the path
-    # The original ordering could be preserved using a special function but
-    # will increase runtime
-    all_uniq_leaf_paths_features = [
-        np.unique(feature_path) for feature_path in all_leaf_paths_features]
-
-    # get the validation classification metrics for the
-    # decision tree against the test data
-    validation_metrics = get_validation_metrics(inp_class_reg_obj=dtree,
-                                                y_true=y_test,
-                                                X_test=X_test)
-
-    # Dictionary of all tree values
-    tree_data = {"num_features_used": num_features_used,
-                 "node_features_idx": node_features_idx,
-                 "max_node_depth": max_node_depth,
-                 "n_nodes": n_nodes,
-                 "all_leaf_node_paths": all_leaf_node_paths,
-                 "all_leaf_nodes": all_leaf_nodes,
-                 "leaf_nodes_depths": leaf_nodes_depths,
-                 "all_leaf_node_samples": all_leaf_node_samples,
-                 "all_leaf_node_samples_percent":
-                 all_leaf_node_samples_percent,
-                 "all_leaf_node_values": all_leaf_node_values,
-                 "all_scaled_leaf_node_values": all_scaled_leaf_node_values,
-                 "tot_leaf_node_values": tot_leaf_node_values,
-                 "all_leaf_node_classes": all_leaf_node_classes,
-                 "all_leaf_paths_features": all_leaf_paths_features,
-                 "all_uniq_leaf_paths_features": all_uniq_leaf_paths_features,
-                 "validation_metrics": validation_metrics}
-    return tree_data
-
-# Get all RF and decision tree data
-
-
-def get_rf_tree_data(rf, X_train, X_test, y_test):
-    """
-    Get the entire fitted random forest and its decision tree data
-    as a convenient dictionary format
-    """
-
-    # random forest feature importances i.e. next iRF iteration weights
-    feature_importances = rf.feature_importances_
-
-    # standard deviation of the feature importances
-    feature_importances_std = np.std(
-        [dtree.feature_importances_ for dtree in rf.estimators_], axis=0)
-    feature_importances_rank_idx = np.argsort(feature_importances)[::-1]
-
-    # get all the validation rf_metrics
-    rf_validation_metrics = get_validation_metrics(inp_class_reg_obj=rf,
-                                                   y_true=y_test,
-                                                   X_test=X_test)
-
-    # Create a dictionary with all random forest metrics
-    # This currently includes the entire random forest fitted object
-    all_rf_tree_outputs = {"rf_obj": rf,
-                           "get_params": rf.get_params,
-                           "rf_validation_metrics": rf_validation_metrics,
-                           "feature_importances": feature_importances,
-                           "feature_importances_std": feature_importances_std,
-                           "feature_importances_rank_idx":
-                           feature_importances_rank_idx}
-
-    # CHECK: Ask SVW if the following should be paralellized!
-    for idx, dtree in enumerate(rf.estimators_):
-        dtree_out = get_tree_data(X_train=X_train,
-                                  X_test=X_test,
-                                  y_test=y_test,
-                                  dtree=dtree,
-                                  root_node_id=0)
-
-        # Append output to our combined random forest outputs dict
-        all_rf_tree_outputs["dtree{}".format(idx)] = dtree_out
-
-    return all_rf_tree_outputs
 
 
 # Random Intersection Tree (RIT)
@@ -632,7 +195,7 @@ def generate_rit_samples(all_rf_tree_data, bin_class_type=1):
     """
 
     # Number of decision trees
-    n_estimators = all_rf_tree_data['rf_obj'].n_estimators
+    n_estimators = all_rf_tree_data['get_params']['n_estimators']
 
     all_weights = []
     all_paths = []
@@ -976,64 +539,41 @@ def run_iRF(X_train,
 
     # Initialize dictionary of bootstrap RIT output
     all_rit_bootstrap_output = {}
-
-    for k in range(K):
-        if k == 0:
-
-            # Initially feature weights are None
-            feature_importances = initial_weights
-
-            # Update the dictionary of all our RF weights
-            all_rf_weights["rf_weight{}".format(k)] = feature_importances
-
-            # fit the classifier
-            rf.fit(
-                X=X_train,
-                y=y_train,
-                feature_weight=all_rf_weights["rf_weight{}".format(k)])
-
-            # Update feature weights using the
-            # new feature importance score
-            feature_importances = rf.feature_importances_
-
-            # Load the weights for the next iteration
-            all_rf_weights["rf_weight{}".format(k + 1)] = feature_importances
-
-        else:
-            # fit weighted RF
-            # Use the weights from the previous iteration
-            rf.fit(
-                X=X_train,
-                y=y_train,
-                feature_weight=all_rf_weights["rf_weight{}".format(k)])
-
-            # Update feature weights using the
-            # new feature importance score
-            feature_importances = rf.feature_importances_
-
-            # Load the weights for the next iteration
-            all_rf_weights["rf_weight{}".format(k + 1)] = feature_importances
-
-        all_K_iter_rf_data["rf_iter{}".format(k)] = get_rf_tree_data(
-            rf=rf,
-            X_train=X_train,
-            X_test=X_test,
-            y_test=y_test)
+    
+    
+    if type(rf) is RandomForestClassifierWithWeights:
+        weightedRF = wrf(**rf.get_params())
+    elif type(rf) is RandomForestRegressorWithWeights:
+        weightedRF = wrf_reg(**rf.get_params())
+    else:
+        raise ValueError('the type of rf cannot be {}'.format(type(rf)))
+    
+    weightedRF.fit(X=X_train, y=y_train, feature_weight = initial_weights, K=K,
+                   X_test = X_test, y_test = y_test)
+    all_rf_weights = weightedRF.all_rf_weights
+    all_K_iter_rf_data = weightedRF.all_K_iter_rf_data
+    
 
     # Run the RITs
     for b in range(B):
 
         # Take a bootstrap sample from the training data
         # based on the specified user proportion
-        X_train_rsmpl, y_rsmpl = resample(
-            X_train, y_train, n_samples=n_samples)
-        # FIXME in iRF R package, when y is discrete, this should be a stratified bootstrap
+        if isinstance(rf, ClassifierMixin):
+            X_train_rsmpl, y_rsmpl = resample(
+                X_train, y_train, n_samples=n_samples, stratify = y_train)
+        else:
+            X_train_rsmpl, y_rsmpl = resample(
+                X_train, y_train, n_samples=n_samples)
+            
+        
 
         # Set up the weighted random forest
         # Using the weight from the (K-1)th iteration i.e. RF(w(K))
-        rf_bootstrap = RandomForestClassifierWithWeights(
-            # CHECK: different number of trees to fit for bootstrap samples
-            n_estimators=n_estimators_bootstrap)
+        rf_bootstrap = clone(rf)
+        
+        # CHECK: different number of trees to fit for bootstrap samples
+        rf_bootstrap.n_estimators=n_estimators_bootstrap
 
         # Fit RF(w(K)) on the bootstrapped dataset
         rf_bootstrap.fit(
